Changes to Introduction, Results, and Discussion (Tim Wang).
For the introduction, the board question was changed to the specific question.

For the results section, precision and recall was added. Accuracy, precision, and recall was organized into a Tibble.

For the discussion section, I discussed results of classification (the plot) and evacuated the classifier. Accuracy was thoroughly discussed, but report still needs a brief discussion of precision and recall. Answering the original question can also be improved.